requirements.txt
# TODO: Update with exact module version
numpy
torch>=1.9
opencv_python
loguru
scikit-image
tqdm
torchvision
Pillow
thop
ninja
tabulate
tensorboard
cython
pycocotools
imagesize

# verified versions
onnx==1.8.1
onnxruntime==1.8.0
onnx-simplifier==0.3.5